Automatically create arduino libraries symbolic links
