+++ sketches/README.md
@@ -18,3 +18,7 @@ Run
 ```
 rosrun rosserial_python serial_node.py _baud:=57600 _port:=<your_m5stack_device>
 ```
+
+# NOTE
+
+The directory `sketch/**/symlink_libs` is automatically created by `catkin build` and the files in the directory is ignored by git. Do not place your original header files here.
